Move Button component to common
